toolkit/components/url-classifier/tests/mochitest/test_lookup_system_principal.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that lookup() on a system principal doesn't crash</title>
  <script type="text/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>

<body>
<div id="content" style="display: none">
</div>
<pre id="test">

<script type="text/javascript">

var Cc = Components.classes;
var Ci = Components.interfaces;

var dbService = Cc["@mozilla.org/url-classifier/dbservice;1"]
                .getService(Ci.nsIUrlClassifierDBService);

dbService.lookup(document.nodePrincipal, "", function(arg) {});

ok(true, "lookup() didn't crash");

</script>
</pre>
</body>
</html>
